[Fix #220] changed namespace to DatarocketsStyle
