--- lib/datarockets/style.rb
@@ -1,14 +1,0 @@
-require "rubocop"
-require "datarockets/style/formatter/todo_list_formatter"
-
-require "datarockets/style/version"
-
-require "datarockets/style/cop/layout/array_alignment_extended"
-require "datarockets/style/cop/style/nested_interpolation"
-
-module Datarockets
-  # Datarickors sharable config
-  module Style
-    # Your code goes here...
-  end
-end
