--- lib/datarockets/style/version.rb
@@ -1,5 +1,0 @@
-module Datarockets
-  module Style
-    VERSION = "1.0.0".freeze
-  end
-end
