+++ lib/datarockets_style.rb
@@ -1,0 +1,12 @@
+require "rubocop"
+require "datarockets_style/formatter/todo_list_formatter"
+
+require "datarockets_style/version"
+
+require "datarockets_style/cop/layout/array_alignment_extended"
+require "datarockets_style/cop/style/nested_interpolation"
+
+# Top level module for datarockets-style
+module DatarocketsStyle
+  # Datarickors sharable config
+end
