+++ lib/datarockets_style/version.rb
@@ -1,0 +1,3 @@
+module DatarocketsStyle
+  VERSION = "1.0.0".freeze
+end
