--- spec/datarockets/style/version_spec.rb
@@ -1,7 +1,0 @@
-RSpec.describe "Datarockets::Style::Version" do
-  describe "::VERSION" do
-    it "contains the current version" do
-      expect(Datarockets::Style::VERSION).not_to be nil
-    end
-  end
-end
