+++ spec/datarockets_style/formatter/todo_list_formatter_spec.rb
